# MIT-licensed SparseMatrixCSC transposition methods