Followup to #13001. This pull request replaces the LGPL-licensed SparseMatrixCSC transposition methods in base/sparse/csparse.jl ([c|f]transpose[!]) with MIT-licensed versions.

Specifically, these methods implement the HALFPERM algorithm described in F. Gustavson, "Two fast algorithms for sparse matrices: multiplication and permuted transposition," ACM TOMS 4(3), 250-269 (1978). The algorithm runs in `O(A.m, A.n, nnz(A))` time and requires no space beyond that passed in. If you know of a faster-in-practice algorithm, please let me know. The new methods' performance is nearly identical to the old methods' performance; the new methods might have a slight edge.

When editing base/sparse/csparse.jl to remove the existing methods, I folded all blocks to avoid peaking at the LGPL code. So someone should check that the removal was graceful given it was blind but for method signatures.

@dmbates (or anyone interested) From what I see when I grep base/sparse/csparse.jl for long-form method signatures, I'm guessing that `triu!` and `tril!` are short children of `fkeep!`; is that so? Are there short-form children of `fkeep!` that I'm not seeing? (I've avoided grepping for short-form methods in case doing so reveals some implementation detail that I should not see.) Additionally, I'm guessing that `f` takes an element (and its position?) and determines whether that element should be kept; what is the implicit signature of `f`? What is the purpose and form of `other`? I might pick off the preceding set of methods next. Concerning `sparse`, I see multiple possible space/time complexity tradeoffs. How much storage does the existing `sparse` method allocate, and does it achieve `O(A.m, A.n, nnz(A))` time or does it trade time to save space? Thanks, and best!