@ViralBShah 2016-01-11
Could you provide some benchmarks just to ensure no performance regressions? 

Cc @jrevels