@Sacha0 2016-01-11
Thanks for the line-length feedback @tkelman! Sorry for somehow nixing the related comments. PR updated accordingly.