@dmbates 2016-01-12
@Sacha0 The signature of `f` in a call to `fkeep!` is
```julia
f{Tv,Ti}(i::Ti, j::Ti, x::Tv, other::Any) → Bool
```

The purpose of the `other` argument is to pass information from the call to `fkeep!` through to the call to `f`.   As you suspect `triu!` and `tril!` are short children of `fkeep!`.  In those cases the `other` argument is the diagonal below which or above which to zero the contents.  It defaults to zero, indicating the main diagonal.

Another use of the `other` argument is in the `droptol!` function where it specifies the tolerance on the absolute value of the non-zeros below which they are dropped.

For details on the `sparse` function I think @ViralBShah is a better source than I.