@Sacha0 2016-01-13
Thanks @dmbates! I will have a go at those methods (`fkeep!`/`triu!`/`tril!`/`droptol!`) next.