@KristofferC 2016-01-14
Would be nice to add some of these benchmarks to BaseBenchmarks.jl